Convert response to TS
