+++ src/utils.ts
@@ -1,10 +1,17 @@
 // eslint-disable-next-line no-undef
-let _process: NodeJS.Process, getNanoSeconds: (() => number) | undefined, loadTime: number | undefined
-// eslint-disable-next-line no-eval, no-empty
-try { _process = eval('typeof __TEST_WEB__ === "undefined" && typeof process === "object" ? process : undefined') } catch (e) {}
+let _process: NodeJS.Process,
+  getNanoSeconds: (() => number) | undefined,
+  loadTime: number | undefined
+
+try {
+  // eslint-disable-next-line no-eval
+  _process = eval(
+    'typeof __TEST_WEB__ === "undefined" && typeof process === "object" ? process : undefined'
+  )
+} catch (e) {} // eslint-disable-line no-empty
 
 const hasProcessHrtime = () => {
-  return (typeof _process !== 'undefined' && _process !== null) && _process.hrtime
+  return typeof _process !== 'undefined' && _process !== null && _process.hrtime
 }
 
 if (hasProcessHrtime()) {
@@ -20,16 +27,17 @@ const R20 = /%20/g
 const isNeitherNullNorUndefined = <T>(x: T | undefined | null): x is T =>
   x !== null && x !== undefined
 
-export const validKeys = (entry: Record<string, unknown>) => Object
-  .keys(entry)
-  .filter((key) => isNeitherNullNorUndefined(entry[key]))
+export const validKeys = (entry: Record<string, unknown>) =>
+  Object.keys(entry).filter((key) => isNeitherNullNorUndefined(entry[key]))
 
 export type Primitive = string | number | boolean
-export const buildRecursive = (key: string, value: Primitive | Primitive[] | Record<string, Primitive>, suffix = ''): string => {
+export const buildRecursive = (
+  key: string,
+  value: Primitive | Primitive[] | Record<string, Primitive>,
+  suffix = ''
+): string => {
   if (Array.isArray(value)) {
-    return value
-      .map((v) => buildRecursive(key, v, suffix + '[]'))
-      .join('&')
+    return value.map((v) => buildRecursive(key, v, suffix + '[]')).join('&')
   }
 
   if (typeof value !== 'object') {
@@ -95,33 +103,36 @@ export const parseResponseHeaders = (headerStr: string) => {
   return headers
 }
 
-export const lowerCaseObjectKeys = (obj: Record<string, unknown>) => {
-  return Object
-    .keys(obj)
-    .reduce((target, key) => {
-      target[key.toLowerCase()] = obj[key]
-      return target
-    }, {} as Record<string, unknown>)
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
+export const lowerCaseObjectKeys = <T extends Record<string, any>>(obj: T) => {
+  return Object.keys(obj).reduce((target, key) => {
+    target[key.toLowerCase() as keyof T] = obj[key]
+    return target
+  }, {} as T)
 }
 
 const hasOwnProperty = Object.prototype.hasOwnProperty
-export const assign = Object.assign || function (target: Record<string, unknown>) {
-  for (let i = 1; i < arguments.length; i++) {
-    // eslint-disable-next-line prefer-rest-params
-    const source = arguments[i]
-    for (const key in source) {
-      if (hasOwnProperty.call(source, key)) {
-        target[key] = source[key]
+export const assign =
+  Object.assign ||
+  function (target: Record<string, unknown>) {
+    for (let i = 1; i < arguments.length; i++) {
+      // eslint-disable-next-line prefer-rest-params
+      const source = arguments[i]
+      for (const key in source) {
+        if (hasOwnProperty.call(source, key)) {
+          target[key] = source[key]
+        }
       }
     }
+    return target
   }
-  return target
-}
 
 const toString = Object.prototype.toString
 export const isPlainObject = (value: unknown): value is Record<string, unknown> => {
-  return toString.call(value) === '[object Object]' &&
+  return (
+    toString.call(value) === '[object Object]' &&
     Object.getPrototypeOf(value) === Object.getPrototypeOf({})
+  )
 }
 
 /**
@@ -138,15 +149,17 @@ export const btoa = (input: string) => {
     // if the next str index does not exist:
     //   change the mapping table to "="
     //   check if d has no fractional digits
-    str.charAt(idx | 0) || (map = '=', idx % 1);
+    str.charAt(idx | 0) || ((map = '='), idx % 1);
     // "8 - idx % 1 * 8" generates the sequence 2, 4, 6, 8
-    output += map.charAt(63 & block >> 8 - idx % 1 * 8)
+    output += map.charAt(63 & (block >> (8 - (idx % 1) * 8)))
   ) {
-    charCode = str.charCodeAt(idx += 3 / 4)
-    if (charCode > 0xFF) {
-      throw new Error("[Mappersmith] 'btoa' failed: The string to be encoded contains characters outside of the Latin1 range.")
+    charCode = str.charCodeAt((idx += 3 / 4))
+    if (charCode > 0xff) {
+      throw new Error(
+        "[Mappersmith] 'btoa' failed: The string to be encoded contains characters outside of the Latin1 range."
+      )
     }
-    block = block << 8 | charCode
+    block = (block << 8) | charCode
   }
   return output
 }
