+++ tsconfig.json
@@ -18,6 +18,6 @@
   ],
   "exclude": [
     "node_modules",
-    "spec/**/*",
+    "spec/typescript/*"
   ]
 }
